Fix large withdrawals
Only the change in src/txs/withdraw.ts is material here. The other
changes were arrived at while debugging and are just good to keep for
consistency.

buildInitiateWithdrawal isn't including the `chain` property in its
output, so we are bypassing it and adding `chain` directly to
`initateWithdrawal` so that it's using the balance on the right chain.
This is definitely a bug in viem.